# api/responses/MemoryMapResponses/keyValueResponse.go
package MemoryMapResponses

type KeyValue struct {
	Key   string `json:"key"`
	Value string `json:"value"`
}

func CreateKeyValueResponse(key, value string) KeyValue {
	return KeyValue{Key: key, Value: value}
}